Added the Arrow Wars Minigame
Added the beginning of the Arrow Wars minigame (not accessible yet), added a message in tchat when a player falls off the map in Spleef, fixed a floating point displayed when it should not be displayed in spleef, and added the spect and spec aliases for /spectate in Spleef

+++ Spleef plugin/src/fr/enderstevegamer/spleef/loops/ActionBarDisplay.java
@@ -30,7 +30,7 @@ public class ActionBarDisplay extends BukkitRunnable {
                         SpleefUtils.sendActionBar(player, ChatColor.GREEN + "You are ready!");
                     }
                     else {
-                        SpleefUtils.sendActionBar(player, ChatColor.GREEN + "You are ready! " + ChatColor.GOLD + "|" + ChatColor.GREEN + " Starting in " + ChatColor.GOLD + SpleefUtils.round(Main.getStartTimer(), 0));
+                        SpleefUtils.sendActionBar(player, ChatColor.GREEN + "You are ready! " + ChatColor.GOLD + "|" + ChatColor.GREEN + " Starting in " + ChatColor.GOLD + (int) SpleefUtils.round(Main.getStartTimer(), 0));
                     }
                 }
                 else if (!Main.getPlayersSpectating().contains(player.getUniqueId())) {
@@ -38,7 +38,7 @@ public class ActionBarDisplay extends BukkitRunnable {
                         SpleefUtils.sendActionBar(player, ChatColor.RED + "You are not ready!");
                     }
                     else {
-                        SpleefUtils.sendActionBar(player, ChatColor.RED + "You are not ready! " + ChatColor.GOLD + "|" + ChatColor.RED + " Starting in " + ChatColor.GOLD + SpleefUtils.round(Main.getStartTimer(), 0));
+                        SpleefUtils.sendActionBar(player, ChatColor.RED + "You are not ready! " + ChatColor.GOLD + "|" + ChatColor.RED + " Starting in " + ChatColor.GOLD + (int) SpleefUtils.round(Main.getStartTimer(), 0));
                     }
                 }
                 else {
@@ -46,7 +46,7 @@ public class ActionBarDisplay extends BukkitRunnable {
                         SpleefUtils.sendActionBar(player, ChatColor.GRAY + "You are spectating!");
                     }
                     else {
-                        SpleefUtils.sendActionBar(player, ChatColor.GRAY + "You are spectating! " + ChatColor.GOLD + "|" + ChatColor.GRAY + " Starting in " + ChatColor.GOLD + SpleefUtils.round(Main.getStartTimer(), 0));
+                        SpleefUtils.sendActionBar(player, ChatColor.GRAY + "You are spectating! " + ChatColor.GOLD + "|" + ChatColor.GRAY + " Starting in " + ChatColor.GOLD + (int) SpleefUtils.round(Main.getStartTimer(), 0));
                     }
                 }
             }
